[DNS] Refactor MediaCapabilitiesCache as interface
Bug: 433773029
