+++ starboard/android/shared/BUILD.gn
@@ -216,8 +216,9 @@ static_library("starboard_platform") {
     "log_raw.cc",
     "max_media_codec_output_buffers_lookup_table.cc",
     "max_media_codec_output_buffers_lookup_table.h",
-    "media_capabilities_cache.cc",
     "media_capabilities_cache.h",
+    "media_capabilities_cache_singleton.cc",
+    "media_capabilities_cache_singleton.h",
     "media_codec_bridge.cc",
     "media_codec_bridge.h",
     "media_common.h",
