+++ starboard/android/shared/media_capabilities_cache.h
@@ -31,160 +31,57 @@
 
 namespace starboard::android::shared {
 
-// TODO: encapsulate a common Range class.
-struct Range {
-  Range() : minimum(0), maximum(0) {}
-  Range(int min, int max) : minimum(min), maximum(max) {}
-  int minimum;
-  int maximum;
-
-  bool Contains(int val) const { return val >= minimum && val <= maximum; }
-};
-
-class CodecCapability {
- public:
-  CodecCapability(JNIEnv* env,
-                  base::android::ScopedJavaLocalRef<jobject>& j_codec_info);
-  virtual ~CodecCapability() {}
-
-  const std::string& name() const { return name_; }
-  bool is_secure_required() const { return is_secure_required_; }
-  bool is_secure_supported() const { return is_secure_supported_; }
-  bool is_tunnel_mode_required() const { return is_tunnel_mode_required_; }
-  bool is_tunnel_mode_supported() const { return is_tunnel_mode_supported_; }
-
- private:
-  CodecCapability(const CodecCapability&) = delete;
-  CodecCapability& operator=(const CodecCapability&) = delete;
-
-  const std::string name_;
-  const bool is_secure_required_;
-  const bool is_secure_supported_;
-  const bool is_tunnel_mode_required_;
-  const bool is_tunnel_mode_supported_;
-};
-
-class AudioCodecCapability : public CodecCapability {
- public:
-  AudioCodecCapability(
-      JNIEnv* env,
-      base::android::ScopedJavaLocalRef<jobject>& j_codec_info,
-      base::android::ScopedJavaLocalRef<jobject>& j_audio_capabilities);
-  ~AudioCodecCapability() override {}
-
-  bool IsBitrateSupported(int bitrate) const;
-
- private:
-  AudioCodecCapability(const AudioCodecCapability&) = delete;
-  AudioCodecCapability& operator=(const AudioCodecCapability&) = delete;
-
-  Range supported_bitrates_;
-};
-
-class VideoCodecCapability : public CodecCapability {
- public:
-  VideoCodecCapability(
-      JNIEnv* env,
-      base::android::ScopedJavaLocalRef<jobject>& j_codec_info,
-      base::android::ScopedJavaLocalRef<jobject>& j_video_capabilities);
-  ~VideoCodecCapability() override;
-
-  bool is_software_decoder() const { return is_software_decoder_; }
-  bool is_hdr_capable() const { return is_hdr_capable_; }
-
-  bool IsBitrateSupported(int bitrate) const;
-  // VideoCodecCapability caches java object MediaCodecInfo.VideoCapabilities.
-  // If improved support check is used,
-  // VideoCapabilities.areSizeAndRateSupported() or
-  // VideoCapabilities.isSizeSupported() will be used to check the
-  // supportability.
-  bool AreResolutionAndRateSupported(int frame_width,
-                                     int frame_height,
-                                     int fps);
-
- private:
-  VideoCodecCapability(const VideoCodecCapability&) = delete;
-  VideoCodecCapability& operator=(const VideoCodecCapability&) = delete;
-
-  const bool is_software_decoder_;
-  const bool is_hdr_capable_;
-  const base::android::ScopedJavaGlobalRef<jobject> j_video_capabilities_;
-  Range supported_widths_;
-  Range supported_heights_;
-  Range supported_bitrates_;
-  Range supported_frame_rates_;
-};
-
 class MediaCapabilitiesCache {
  public:
-  static MediaCapabilitiesCache* GetInstance();
+  virtual ~MediaCapabilitiesCache() = default;
 
-  bool IsWidevineSupported();
-  bool IsCbcsSchemeSupported();
+  static MediaCapabilitiesCache* GetInstance() { return nullptr; }
 
-  bool IsHDRTransferCharacteristicsSupported(SbMediaTransferId transfer_id);
+  virtual bool IsWidevineSupported() = 0;
+  virtual bool IsCbcsSchemeSupported() = 0;
 
-  bool IsPassthroughSupported(SbMediaAudioCodec codec);
+  virtual bool IsHDRTransferCharacteristicsSupported(
+      SbMediaTransferId transfer_id) = 0;
 
-  bool GetAudioConfiguration(int index,
-                             SbMediaAudioConfiguration* configuration);
+  virtual bool IsPassthroughSupported(SbMediaAudioCodec codec) = 0;
 
-  bool HasAudioDecoderFor(const std::string& mime_type, int bitrate);
+  virtual bool GetAudioConfiguration(
+      int index,
+      SbMediaAudioConfiguration* configuration) = 0;
 
-  bool HasVideoDecoderFor(const std::string& mime_type,
-                          bool must_support_secure,
-                          bool must_support_hdr,
-                          bool must_support_tunnel_mode,
-                          int frame_width,
-                          int frame_height,
-                          int bitrate,
-                          int fps);
+  virtual bool HasAudioDecoderFor(const std::string& mime_type,
+                                  int bitrate) = 0;
 
-  std::string FindAudioDecoder(const std::string& mime_type, int bitrate);
+  virtual bool HasVideoDecoderFor(const std::string& mime_type,
+                                  bool must_support_secure,
+                                  bool must_support_hdr,
+                                  bool must_support_tunnel_mode,
+                                  int frame_width,
+                                  int frame_height,
+                                  int bitrate,
+                                  int fps) = 0;
 
-  std::string FindVideoDecoder(const std::string& mime_type,
-                               bool must_support_secure,
-                               bool must_support_hdr,
-                               bool require_software_codec,
-                               bool must_support_tunnel_mode,
-                               int frame_width,
-                               int frame_height,
-                               int bitrate,
-                               int fps);
+  virtual std::string FindAudioDecoder(const std::string& mime_type,
+                                       int bitrate) = 0;
 
-  bool IsEnabled() const { return is_enabled_; }
-  void SetCacheEnabled(bool enabled) { is_enabled_ = enabled; }
-  void ClearCache() { capabilities_is_dirty_ = true; }
+  virtual std::string FindVideoDecoder(const std::string& mime_type,
+                                       bool must_support_secure,
+                                       bool must_support_hdr,
+                                       bool require_software_codec,
+                                       bool must_support_tunnel_mode,
+                                       int frame_width,
+                                       int frame_height,
+                                       int bitrate,
+                                       int fps) = 0;
 
- private:
-  MediaCapabilitiesCache();
-  ~MediaCapabilitiesCache() {}
+  virtual bool IsEnabled() const = 0;
+  virtual void SetCacheEnabled(bool enabled) = 0;
+  virtual void ClearCache() = 0;
 
+ protected:
+  MediaCapabilitiesCache() = default;
   MediaCapabilitiesCache(const MediaCapabilitiesCache&) = delete;
   MediaCapabilitiesCache& operator=(const MediaCapabilitiesCache&) = delete;
-
-  void UpdateMediaCapabilities_Locked();
-  void LoadAudioConfigurations_Locked();
-  void LoadCodecInfos_Locked();
-
-  std::mutex mutex_;
-
-  std::set<SbMediaTransferId> supported_transfer_ids_;
-  std::map<SbMediaAudioCodec, bool> passthrough_supportabilities_;
-
-  typedef std::vector<std::unique_ptr<AudioCodecCapability>>
-      AudioCodecCapabilities;
-  typedef std::vector<std::unique_ptr<VideoCodecCapability>>
-      VideoCodecCapabilities;
-
-  std::map<std::string, AudioCodecCapabilities> audio_codec_capabilities_map_;
-  std::map<std::string, VideoCodecCapabilities> video_codec_capabilities_map_;
-  std::vector<SbMediaAudioConfiguration> audio_configurations_;
-  bool is_widevine_supported_ = false;
-  bool is_cbcs_supported_ = false;
-
-  std::atomic_bool is_enabled_{true};
-  std::atomic_bool capabilities_is_dirty_{true};
 };
 
 }  // namespace starboard::android::shared
